Define Registry.__rmatmul__ for adding resources.
I can't resist a bit of over-the-top fun. This is a (useful) shortcut
for adding a resource (or multiple resources) to a registry where the
URI you want to use is specifically the resource's internal ID.

It *must* have one to use this, otherwise you get an error, which is a
benefit over using `with_resource` -- which will happily not error if
you pass it `resource.id()` but that returned None.

That being said, Registries already have 3 methods named with_<foo> for
adding resource-like things to themselves, so I didn't want to add a
fourth and risk users getting even more confused as to which they should
use.

Using @ seems like the sort of thing that "pOwEr UsErS" will find, and
more casual users who just want to figure out something working will not
be distracted by.